+++ internal/gatewayapi/translator.go
@@ -3,7 +3,6 @@ package gatewayapi
 import (
 	"fmt"
 	"net/netip"
-	"sort"
 	"strings"
 
 	"golang.org/x/exp/slices"
@@ -125,7 +124,7 @@ func (t *Translator) Translate(resources *Resources) *TranslateResult {
 	httpRoutes := t.ProcessHTTPRoutes(resources.HTTPRoutes, gateways, resources, xdsIR)
 
 	// Sort xdsIR based on the Gateway API spec
-	t.sortXdsIR(xdsIR)
+	sortXdsIRMap(xdsIR)
 
 	return newTranslateResult(gateways, httpRoutes, xdsIR, infraIR)
 }
@@ -1132,63 +1131,6 @@ func (t *Translator) ProcessHTTPRoutes(httpRoutes []*v1beta1.HTTPRoute, gateways
 	}
 
 	return relevantHTTPRoutes
-}
-
-// sortXdsIR sorts the xdsIR based on the match precedence
-// defined in the Gateway API spec.
-// https://gateway-api.sigs.k8s.io/references/spec/#gateway.networking.k8s.io/v1beta1.HTTPRouteRule
-func (t *Translator) sortXdsIR(xdsIR XdsIRMap) {
-	for _, ir := range xdsIR {
-		ir := ir
-		sort.SliceStable(ir.HTTP, func(i, j int) bool { return ir.HTTP[i].Name < ir.HTTP[j].Name })
-		for _, http := range ir.HTTP {
-			http := http
-			sort.Slice(http.Routes, func(i, j int) bool {
-				// 1. Sort based on characters in a matching path.
-				pCountI := pathMatchCount(http.Routes[i].PathMatch)
-				pCountJ := pathMatchCount(http.Routes[j].PathMatch)
-				if pCountI < pCountJ {
-					return true
-				}
-				if pCountI > pCountJ {
-					return false
-				}
-				// Equal case
-
-				// 2. Sort based on the number of Header matches.
-				hCountI := len(http.Routes[i].HeaderMatches)
-				hCountJ := len(http.Routes[j].HeaderMatches)
-				if hCountI < hCountJ {
-					return true
-				}
-				if hCountI > hCountJ {
-					return false
-				}
-				// Equal case
-
-				// 3. Sort based on the number of Query param matches.
-				qCountI := len(http.Routes[i].QueryParamMatches)
-				qCountJ := len(http.Routes[j].QueryParamMatches)
-				return qCountI < qCountJ
-			})
-		}
-	}
-}
-
-func pathMatchCount(pathMatch *ir.StringMatch) int {
-	var count int
-	if pathMatch != nil {
-		if pathMatch.Exact != nil {
-			count += len(*pathMatch.Exact)
-		}
-		if pathMatch.Prefix != nil {
-			count += len(*pathMatch.Prefix)
-		}
-		if pathMatch.SafeRegex != nil {
-			count += len(*pathMatch.SafeRegex)
-		}
-	}
-	return count
 }
 
 type crossNamespaceFrom struct {
